RAG/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
html2text
unstructured
qdrant-client